+++ bake/azext_bake/commands.py
@@ -3,8 +3,8 @@
 # Licensed under the MIT License.
 # ------------------------------------
 
-from ._validators import (builder_validator, process_bake_image_namespace,
-                          process_bake_repo_namespace,
+from ._validators import builder_validator  # process_bake_image_namespace,
+from ._validators import (process_bake_repo_namespace,
                           process_bake_repo_validate_namespace,
                           process_sandbox_create_namespace)
 
@@ -15,9 +15,10 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         pass
 
     with self.command_group('bake') as g:
+        g.custom_command('version', 'bake_version')
         g.custom_command('upgrade', 'bake_upgrade')
         g.custom_command('repo', 'bake_repo', validator=process_bake_repo_namespace)
-        g.custom_command('image', 'bake_image', validator=process_bake_image_namespace)
+        # g.custom_command('image', 'bake_image', validator=process_bake_image_namespace)
 
     with self.command_group('bake sandbox') as g:
         g.custom_command('create', 'bake_sandbox_create', validator=process_sandbox_create_namespace)
@@ -30,16 +31,11 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         g.custom_command('export', 'bake_yaml_export')
         g.custom_command('validate', 'bake_yaml_validate')
 
-    # with self.command_group('bake image') as g:
-    #     g.custom_command('test', 'bake_image_test')
+    with self.command_group('bake validate') as g:
+        # g.custom_command('image', 'bake_validate_image')
+        g.custom_command('repo', 'bake_repo_validate', validator=process_bake_repo_validate_namespace)
+        g.custom_command('sandbox', 'bake_sandbox_validate')
+        g.custom_command('yaml', 'bake_yaml_validate')
 
     with self.command_group('bake _builder') as g:
         g.custom_command('build', 'bake_builder_build', validator=builder_validator)
-
-    # with self.command_group('bake validate') as g:
-    #     g.custom_command('image', 'bake_validate_image')
-    #     g.custom_command('repo', 'bake_validate_repo', validator=process_bake_repo_validate_namespace)
-    #     g.custom_command('sandbox', 'bake_validate_sandbox')
-    #     g.custom_command('yaml', 'bake_validate_yaml')
-
-    # with self.command_group('bake')