+++ bake/azext_bake/custom.py
@@ -8,15 +8,18 @@ import json
 import os
 
 import yaml
+from azure.cli.core.extension.operations import (show_extension,
+                                                 update_extension)
 from knack.util import CLIError
+from packaging.version import parse
 
 from ._arm import (create_image_definition,
                    deploy_arm_template_at_resource_group,
                    ensure_gallery_permissions, get_arm_output, get_gallery,
                    get_image_definition, image_version_exists)
 from ._constants import IN_BUILDER
-from ._github import (get_github_release, get_release_templates,
-                      get_template_url)
+from ._github import (get_github_latest_release_version, get_github_release,
+                      get_release_templates, get_template_url)
 from ._packer import (check_packer_install, copy_packer_files,
                       inject_choco_provisioners, inject_winget_provisioners,
                       packer_execute, save_packer_vars_file)
@@ -174,7 +177,7 @@ def bake_repo_validate(cmd, repository_path, sandbox=None, gallery=None, images=
     # raise CLIError('Not implemented')
 
 
-def bake_sandbox_create(cmd, location, sandbox_resource_group_name, name_prefix,
+def bake_sandbox_create(cmd, location, sandbox_resource_group_name, name_prefix, gallery_resource_id=None,
                         tags=None, principal_id=None, vnet_address_prefix='10.0.0.0/24',
                         default_subnet_name='default', default_subnet_address_prefix='10.0.0.0/25',
                         builders_subnet_name='builders', builders_subnet_address_prefix='10.0.0.128/25',
@@ -209,17 +212,21 @@ def bake_sandbox_create(cmd, location, sandbox_resource_group_name, name_prefix,
     params.append(f'builderSubnetAddressPrefix={builders_subnet_address_prefix}')
     params.append(f'tags={json.dumps(tags)}')
 
+    if gallery_resource_id:
+        params.append('galleryIds={}'.format(json.dumps([gallery_resource_id])))
+
     hook.add(message='Creating sandbox environment')
     logger.info(f'Deploying sandbox...')
-    deployment, _ = deploy_arm_template_at_resource_group(cmd, sandbox_resource_group_name,
-                                                          template_file=template_file, template_uri=template_uri,
-                                                          parameters=[params])
+    deployment, outputs = deploy_arm_template_at_resource_group(cmd, sandbox_resource_group_name,
+                                                                template_file=template_file, template_uri=template_uri,
+                                                                parameters=[params])
     logger.info(f'Finished deploying sandbox')
     logger.warning(f'Successfully deployed sandbox environment')
     logger.warning(f'You can configure it as your default sandbox using `az configure --defaults bake-sandbox={sandbox_resource_group_name}`')
 
     hook.end(message=' ')
-    return True
+
+    return
 
 
 def bake_yaml_export(cmd, sandbox_resource_group_name, gallery_resource_id, sandbox=None, gallery=None, images=None,
@@ -234,14 +241,17 @@ def bake_yaml_export(cmd, sandbox_resource_group_name, gallery_resource_id, sand
     if images:
         bake_obj['images'] = images
 
+    yaml_schema = '# yaml-language-server: $schema=https://github.com/colbylwilliams/az-bake/releases/latest/download/bake.schema.json\n'
+    yaml_str = yaml_schema + yaml.safe_dump(bake_obj, default_flow_style=False, sort_keys=False)
+
     if stdout:
-        print(yaml.safe_dump(bake_obj, default_flow_style=False, sort_keys=False))
+        print(yaml_str)
     elif outfile:
         with open(outfile, 'w') as f:
-            yaml.safe_dump(bake_obj, f, default_flow_style=False, sort_keys=False)
+            f.write(yaml_str)
     elif outdir:
         with open(outdir / 'bake.yml', 'w') as f:
-            yaml.safe_dump(bake_obj, f, default_flow_style=False, sort_keys=False)
+            f.write(yaml_str)
 
     return
 
@@ -264,17 +274,54 @@ def bake_image(cmd, image_path, sandbox_resource_group_name=None, bake_yaml=None
     raise CLIError('Not implemented')
 
 
+def bake_version(cmd):
+    ext = show_extension('bake')
+    current_version = 'v' + ext['version']
+    logger.info(f'Current version: {current_version}')
+    current_version_parsed = parse(current_version)
+    print(f'az bake version: {current_version}')
+
+    latest_version = get_github_latest_release_version()
+    logger.info(f'Latest version: {latest_version}')
+    latest_version_parsed = parse(latest_version)
+
+    if current_version_parsed < latest_version_parsed:
+        logger.warning(f'There is a new version of az bake {latest_version}. Please update using: az bake upgrade')
+    return
+
+
 def bake_upgrade(cmd, version=None, prerelease=False):
-    from azure.cli.core.extension.operations import update_extension
+    ext = show_extension('bake')
+    current_version = 'v' + ext['version']
+    logger.info(f'Current version: {current_version}')
+    current_version_parsed = parse(current_version)
+
     release = get_github_release(version=version, prerelease=prerelease)
-    index = next((a for a in release['assets']
-                  if 'index.json' in a['browser_download_url']), None)
+
+    new_version = release['tag_name']
+    logger.info(f'Latest{" prerelease" if prerelease else ""} version: {new_version}')
+    new_version_parsed = parse(new_version)
+
+    is_dev = 'extensionType' in ext and ext['extensionType'] == 'dev'
+
+    if not is_dev and new_version_parsed == current_version_parsed:
+        print(f'Already on latest{" prerelease" if prerelease else ""} version: {new_version}')
+        return
+
+    if not is_dev and new_version_parsed < current_version_parsed:
+        print(f'Current version is newer than latest{" prerelease" if prerelease else ""} version: {new_version}')
+        return
+
+    logger.info(f'Upgrading to latest{" prerelease" if prerelease else ""} version: {new_version}')
+    index = next((a for a in release['assets'] if 'index.json' in a['browser_download_url']), None)
 
     index_url = index['browser_download_url'] if index else None
     if not index_url:
-        raise CLIError(
-            f"Could not find index.json asset on release {release['tag_name']}. "
-            'Specify a specific prerelease version with --version '
-            'or use latest prerelease with --pre')
+        raise CLIError(f'Could not find index.json asset on release {new_version}. '
+                       'Specify a specific prerelease version with --version/-v or use latest prerelease with --pre')
 
-    update_extension(cmd, extension_name='bake', index_url=index_url)
+    if is_dev:
+        logger.warning(f'Skipping upgrade of dev extension.')
+        return
+    else:
+        return update_extension(cmd, extension_name='bake', index_url=index_url)